+++ lib/model/product.dart
@@ -1,11 +1,4 @@
 class Product {
-  final String text;
-  final String owner;
-  final String amount;
-  final String image;
-  final String seller;
-  final int height;
-
   Product({
     this.text,
     this.owner,
@@ -14,4 +7,11 @@ class Product {
     this.seller,
     this.height,
   });
+
+  final String text;
+  final String owner;
+  final String amount;
+  final String image;
+  final String seller;
+  final int height;
 }
